# [SPARK-17972][SQL] Cache analyzed plan instead of optimized plan to avoid slow query planning