## What changes were proposed in this pull request?

Iterative ML code may easily create query plans that grow exponentially. We found that query planning time also increases exponentially even when all the sub-plan trees are cached.

The following snippet illustrates the problem:

```scala
(0 until 6).foldLeft(Seq(1, 2, 3).toDS) { (plan, iteration) =>
  val start = System.currentTimeMillis()
  val result = plan.join(plan, "value").join(plan, "value").join(plan, "value").join(plan, "value")
  result.cache()
  System.out.println(s"Iteration $iteration takes time ${System.currentTimeMillis() - start} ms")
  result.as[Int]
}

// Iteration 0 takes time 9 ms
// Iteration 1 takes time 19 ms
// Iteration 2 takes time 61 ms
// Iteration 3 takes time 219 ms
// Iteration 4 takes time 830 ms
// Iteration 5 takes time 4080 ms
```

This is because when caching a Dataset, what is cached is actually the optimized plan, while we always build new Datasets upon analyzed plans of existing Datasets. By caching analyzed plans instead of optimized plans, this issue can be resolved.

Results produced by the same snippet (with more iterations) shown above after applying this PR:

```scala
// Iteration 0 takes time 9 ms
// Iteration 1 takes time 45 ms
// Iteration 2 takes time 42 ms
// Iteration 3 takes time 33 ms
// Iteration 4 takes time 37 ms
// Iteration 5 takes time 64 ms
// Iteration 6 takes time 42 ms
// Iteration 7 takes time 38 ms
// Iteration 8 takes time 50 ms
// Iteration 9 takes time 53 ms
// Iteration 10 takes time 48 ms
// Iteration 11 takes time 45 ms
// Iteration 12 takes time 49 ms
// Iteration 13 takes time 46 ms
// Iteration 14 takes time 45 ms
// Iteration 15 takes time 41 ms
// Iteration 16 takes time 49 ms
// Iteration 17 takes time 46 ms
// Iteration 18 takes time 38 ms
// Iteration 19 takes time 32 ms
```

## How was this patch tested?

Existing tests.