@cloud-fan 2016-10-18
We need to swap `analyzed` and `optimized` in your PR description some where.