@liancheng 2016-10-18
The previous test failure was because we replace the analyzed plan with `withCacheData`, while cache manager uses the original analyzed plan as keys.

Force-pushed a new and much simpler approach by building new Datasets upon `withCacheData`. Let's see whether Jenkins passes.