@yhuai 2016-10-19
cc @mengxr and @jkbradley