@liancheng 2016-10-20
The most recent version still breaks some test cases related to caching. Investigating it.