@naliazheli 2016-10-22
dataset.checkpoint  is what i need.